World Cup/WorldCup/pyproject.toml
[tool.poetry]
name = "worldcup"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
django-debug-toolbar = "^3.8.1"
djangorestframework = "^3.14.0"
windows-curses = "^2.3.1"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
